requirements-dev.txt
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
Faker==30.6.0
ipython==8.28.0
ipdb==0.13.13
